i think we've ignored destination

+++ Rakefile
@@ -1,5 +1,6 @@
 require 'fileutils'
 require 'tmpdir'
+require 'rake'
 
 desc 'Generate deck from Travis CI and publish to GitHub Pages.'
 task :travis do
@@ -15,10 +16,10 @@ task :travis do
   rev = %x(git rev-parse HEAD).strip
 
   Dir.mktmpdir do |dir|
-    dir = File.join dir, 'site'
+    destination = File.join dir, 'site'
     sh 'asciidoctor Main.adoc -o index.html'
     #sh 'asciidoctor Main.adoc -b docbook5'
-    #fail "Build failed." unless Dir.exists? destination
+    fail "Build failed." unless Dir.exists?(destination)
     sh "git clone --branch #{deploy_branch} #{repo} #{dir}"
     sh %Q(rsync -rt --del --exclude=".git" #{destination} #{dir})
     Dir.chdir dir do
